src: refactor milestone tests, add utils, add producer check
