+++ src/express/commands/milestone.js
@@ -4,182 +4,22 @@ import { timer } from '../common/time-utils'
 import { checkLatestMilestone } from './monitor';
 
 const {
-  runSshCommand,
   maxRetries,
   runCommand,
-  runSshCommandWithReturn
 } = require('../common/remote-worker')
+
+import {
+  getBlock,
+  getValidatorInfo,
+  getPeerLength,
+  createClusters,
+  rejoinClusters,
+  getEnode,
+  validateNumberOfPeers
+} from '../common/milestone-utils'
 
 const milestoneLength = 64
 const queryTimer = (milestoneLength / 4) * 1000
-
-export async function getBlock(ip, number = 'latest') {
-  const url = `http://${ip}:8545`
-  if (number != 'latest' && number != 'finalized') {
-    number = '0x' + Number(number).toString(16) // hexify
-  }
-
-  const opts = {
-    'jsonrpc':'2.0',
-    'id':1,
-    'method':'eth_getBlockByNumber',
-    'params':[number, false],
-  }
-
-  const response = await fetch(url, {
-    method: 'POST',
-    headers: {'Content-Type': 'application/json'},
-    body: JSON.stringify(opts)
-  })
-  
-  const responseJson = await response.json()
-  if (responseJson.result) {
-    // console.log(`📍Request. number: ${number}, opts: ${JSON.stringify(opts)}`)
-    return responseJson.result
-  } else {
-    console.log(`📍Error fetching block. number: ${number}, opts: ${JSON.stringify(opts)}`)
-    console.log(`📍Response received:`, responseJson)
-  }
-
-  return undefined
-}
-
-async function removePeers(ip, peers) {
-  let tasks = []
-  for (let i = 0; i < peers.length; i++) {
-    let command = `~/go/bin/bor attach ~/.bor/data/bor.ipc --exec "admin.removePeer('${peers[i]}')"`
-    tasks.push(runSshCommand(ip, command, maxRetries))
-  }
-
-  let response = false
-  await Promise.all(tasks).then(() => {
-    response = true
-  }).catch((error) => {
-    console.log('📍Unable to remove peers, error:', error)
-  })
-
-  return response
-}
-
-export async function addPeers(ip, peers) {
-  let tasks = []
-  for (let i = 0; i < peers.length; i++) {
-    let command = `~/go/bin/bor attach ~/.bor/data/bor.ipc --exec "admin.addPeer('${peers[i]}')"`
-    tasks.push(runSshCommand(ip, command, maxRetries))
-  }
-
-  let response = false
-  await Promise.all(tasks).then(() => {
-    response = true   
-  }).catch((error) => {
-    console.log('📍Unable to add peers, error:', error)
-  })
-
-  return response
-}
-
-export async function getPeerLength(ip) {
-  const command = `/home/ubuntu/go/bin/bor attach ~/.bor/data/bor.ipc --exec "admin.peers.length"`
-  try {
-    let length = await runSshCommandWithReturn(ip, command, maxRetries)
-    return parseInt(length)
-  } catch (error) {
-    console.log('📍Unable to query peer length, error:', error)
-  }
-
-  return -1
-}
-
-async function createClusters(ips, enodes) {
-  let tasks = []
-  for (let i = 0; i < ips.length; i++) {
-    if (i == 0) {
-      // remove all other peers from 1st node
-      tasks.push(removePeers(ips[i], enodes.slice(1)))
-    } else {
-      // remove 1st node from all peers
-      tasks.push(removePeers(ips[i], [enodes[0]]))
-    }
-  }
-
-  let response = false
-  await Promise.all(tasks).then((values) => {
-    if (values.includes(false) || values.includes(undefined)) {
-      console.log('📍Unable to remove peers, exiting')
-    } else {
-      response = true
-    }
-  }).catch(error => {
-    console.log('📍Unable to remove peers, error:', error)
-  })
-
-  return response
-}
-
-async function rejoinClusters(ips, enodes) {
-  let tasks = []
-  for (let i = 0; i < ips.length; i++) {
-    if (i == 0) {
-      // add all other peers in 1st node
-      tasks.push(addPeers(ips[i], enodes.slice(1)))
-    } else {
-      // add 1st node in all peers
-      tasks.push(addPeers(ips[i], [enodes[0]]))
-    }
-  }
-
-  let response = false
-  await Promise.all(tasks).then(values => {
-    if (values.includes(false) || values.includes(undefined)) {
-      console.log('📍Unable to add peers, exiting')
-    } else {
-      response = true
-    }
-  }).catch(error => {
-    console.log("📍Unable to add peers, error", error)
-  })
-  
-  return response
-}
-
-export async function getEnode(user, host) {
-  const ip = `${user}@${host}`
-  let command = '~/go/bin/bor attach ~/.bor/data/bor.ipc --exec admin.nodeInfo.enode'
-  try {
-    const fullEnode = await runSshCommandWithReturn(ip, command, maxRetries)
-    let enode = String(fullEnode).split('@')[0].slice(1); // remove the local ip from the enode
-    if (enode.length != 136) { // prefix "enode://" + 128 hex values for enode itself
-      return ''
-    }
-    enode += "@" + host + ":30303" // assuming that p2p port is opened on 30303 
-    return enode
-  } catch (error) {
-    console.log(`📍Unable to query enode. Error: ${error}`)
-  }
-
-  return ''
-}
-
-function validateNumberOfPeers(peers) {
-  let recreate = false
-
-  // 1st node should have 0 peers
-  if (peers[0] > 0) {
-    console.log('📍Unexpected peer length received for 1st cluster, retrying. expected: 0, received:', peers[0])
-    recreate = true
-  }
-
-  // Remaining nodes should have total-1 peers
-  for (let i = 1; i < peers.length; i++) {
-    if (peers[i] != (peers.length - 2)) {
-      console.log('📍Unexpected peer length received for 2nd cluster, retrying')
-      recreate = true
-      break
-    }
-  }
-
-  return recreate
-}
 
 export async function milestone() {
   // NOTE: Make sure bor branch has logic for hardcoded primary validator
@@ -255,7 +95,7 @@ export async function milestone() {
   console.log('📍Rejoined clusters')
 
   console.log('📍Waiting to fetch finalized blocks...')
-  await timer(10000)
+  // await timer(32000)
 
   // Fetch the last 'finalized' block
   console.log('📍Trying to fetch last finalized block')
@@ -339,11 +179,11 @@ export async function milestone() {
   // Cluster 1 has a single primary producer whose difficulty should always be higher. 
   // Cluster 2 should have remaining nodes (with 2/3+1 stake) all with difficulty lower than node 1
   // and nodes performing mining out of sync. 
-  await timer(10000)
 
   // Validate if both the clusters are on their own chain. 
-  console.log('📍Trying to fetch latest block from both clusters')
-  
+  console.log('📍Trying to fetch latest block from both clusters after 10s')
+  await timer(10000)
+
   // We'll fetch block from cluster 2 first as it'll be behind in terms of block height
   let latestBlockCluster2 = await runCommand(getBlock, borHosts[1], "latest", maxRetries)
   if (latestBlockCluster2 == undefined) {
@@ -407,6 +247,34 @@ export async function milestone() {
 
   let latestMilestone = milestone.result
   console.log(`📍Got milestone from heimdall. Start block: ${Number(latestMilestone.start_block)}, End block: ${Number(latestMilestone.end_block)}, ID: ${latestMilestone.milestone_id}`)
+
+  // Validate if the milestone is proposed by validators of cluster 2 and not by validators of cluster 1
+  let validators = await getValidatorInfo(ips[0])
+  try {
+    if (validators) {
+      if (latestMilestone.proposer == validators[0].address) {
+        console.log(`📍Invalid milestone got proposed from validator/s of cluster 1. Proposer: ${latestMilestone.proposer}, Validators address: ${validators[0].address}, exiting`)
+        return
+      }
+  
+      // Skip the validator from cluster 1
+      let done = false
+      for (let i = 1; i < validators.length; i++) {
+        if (latestMilestone.proposer == validators[i].address) {
+          console.log(`📍Validated milestone proposer`)
+          done = true
+          break
+        }
+      }
+
+      if (!done) {
+        console.log('📍Invalid milestone got proposed from validator/s of cluster 1, proposer: exiting')
+        return
+      }
+    }
+  } catch (error) {
+    console.log('📍Error in validating milestone proposer')
+  }
 
   console.log('📍Waiting for bor nodes to import milestone')
   await timer(32000)
